minor utils refactor
