+++ lib/mixins/browser_actions.js
@@ -2,7 +2,7 @@
 "use strict";
 
 const ErrorFactory = require('../errors/error_factory');
-const utils = require('../utils/utils');
+const utils = require('../utils');
 
 module.exports = function BrowserActionsMixin(s) {
     return class BrowserActions extends s {
