+++ lib/modules/assertions/browser_assertions.js
@@ -1,8 +1,8 @@
 /* global WendigoUtils */
 "use strict";
 
+const utils = require('../../utils');
 const elementsAssertionUtils = require('./utils/assert_elements');
-const textUtils = require('./utils/text_utils');
 const assertUtils = require('./utils/assert_utils');
 const ErrorFactory = require('../../errors/error_factory');
 const BrowserModule = require('../browser_module');
@@ -31,7 +31,7 @@ module.exports = class BrowserAssertions extends BrowserModule {
         if(!Array.isArray(expected)) expected = [expected];
         return this._browser.text(selector).then((texts) => {
             for(const expectedText of expected) {
-                if(!textUtils.matchTextList(texts, expectedText)) {
+                if(!utils.matchTextList(texts, expectedText)) {
                     if(!msg) {
                         const foundText = texts.length === 0 ? "no text" : `"${texts.join(" ")}"`;
                         msg = `Expected element "${selector}" to have text "${expectedText}", ${foundText} found`;
@@ -58,7 +58,7 @@ module.exports = class BrowserAssertions extends BrowserModule {
     title(expected, msg) {
         return this._browser.title().then((title) => {
             const foundTitle = title ? `"${title}"` : "no title";
-            if(!textUtils.matchText(title, expected)) {
+            if(!utils.matchText(title, expected)) {
                 if(!msg) msg = `Expected page title to be "${expected}", ${foundTitle} found`;
                 return assertUtils.rejectAssertion(msg);
             }
@@ -174,7 +174,7 @@ module.exports = class BrowserAssertions extends BrowserModule {
                 return Promise.reject(error);
             }
             for(const html of found) {
-                if(textUtils.matchText(html, expected)) return Promise.resolve();
+                if(utils.matchText(html, expected)) return Promise.resolve();
             }
 
             if(!msg) {
