+++ lib/modules/assertions/browser_not_assertions.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const textUtils = require('./utils/text_utils');
+const utils = require('../../utils');
 const assertUtils = require('./utils/assert_utils');
 
 module.exports = class BrowserNotAssertions {
@@ -30,7 +30,7 @@ module.exports = class BrowserNotAssertions {
         if(!Array.isArray(expected)) expected = [expected];
         return this._browser.text(selector).then((texts) => {
             for(const expectedText of expected) {
-                if(textUtils.matchTextList(texts, expectedText)) {
+                if(utils.matchTextList(texts, expectedText)) {
                     if(!msg) msg = `Expected element "${selector}" not to have text "${expectedText}"`;
                     return assertUtils.rejectAssertion(msg);
                 }
