+++ lib/utils.js
@@ -1,6 +1,14 @@
 "use strict";
 
 module.exports = {
+    isNumber(n) {
+        return !Number.isNaN(Number(n));
+    },
+    promiseSerial(funcs) {
+        return funcs.reduce((promise, func) => {
+            return promise.then(result => func().then(Array.prototype.concat.bind(result)));
+        }, Promise.resolve([]));
+    },
     matchText(text, expected) {
         if(expected instanceof RegExp) {
             return expected.test(text);
@@ -12,5 +20,4 @@ module.exports = {
         }
         return false;
     }
-
 };
