+++ tests/assertions/assert_inner_html.test.js
@@ -2,7 +2,7 @@
 
 const Wendigo = require('../../lib/wendigo');
 const configUrls = require('../config.json').urls;
-const utils = require('../utils');
+const utils = require('../test_utils');
 
 describe("Assert Inner Html", function() {
     this.timeout(5000);
