+++ tests/assertions/assert_local_storage.test.js
@@ -2,7 +2,7 @@
 
 const Wendigo = require('../../lib/wendigo');
 const configUrls = require('../config.json').urls;
-const utils = require('../utils');
+const utils = require('../test_utils');
 
 
 describe("Assert Local Storage", function() {
