+++ tests/browser/click.test.js
@@ -2,8 +2,8 @@
 
 const assert = require('assert');
 const Wendigo = require('../../lib/wendigo');
-const utils = require('.././utils');
-const configUrls = require('.././config.json').urls;
+const utils = require('../test_utils');
+const configUrls = require('../config.json').urls;
 
 describe("Click", function() {
     this.timeout(5000);
