+++ tests/browser/wait_until_not_visible.test.js
@@ -2,7 +2,7 @@
 
 const Wendigo = require('../../lib/wendigo');
 const configUrls = require('../config.json').urls;
-const utils = require('../utils');
+const utils = require('../test_utils');
 
 describe("Wait Until Not Visible", function() {
     this.timeout(5000);
